requirements.txt
selene~=2.0.0rc4
allure-pytest~=2.13.2
pytest